add failing spec for active record transaction support
When using machinist in an active record transaction, rolling back doesn't
completely clean the database. It looks like machinist is doing some work
outside the transaction and thus making changes out of the rollback's
scope. Since it's common to use this 'transactional fixture' technique to
ensure a repeatable db state while testing, this should be fixed.